feat: implement issues #329, #330, #331, #339 with pre-existing bug fixes
Add four major features and fix pre-existing compilation errors that
prevented the project from building (broken async-graphql dependency).

## New features

### #329 D-13: Contract Security Audit Integration
- `starforge security audit` command with Slither/Mythril integration
- Built-in static analysis via existing checklist module
- Security scoring (0-100) with penalty weights per severity level
- Remediation suggestions and JSON/text output formats
- CI-friendly `--min-score` threshold flag

### #330 D-14: Deployment Rollback System
- `starforge deployments history/rollback/verify/dashboard/approve`
- JSON deployment history at ~/.starforge/deploy_history.json
- Generates stellar CLI rollback commands with wasm hash
- On-chain wallet verification and per-network dashboard stats
- Approval workflow for pending deployments

### #331 D-15: Cross-Contract Call Visualization
- `starforge contract call-graph` command
- Extracts invoke_contract!, Client::new, and ::Client patterns
- ASCII tree, Graphviz DOT, and JSON output formats
- Pattern detection: re-entrancy risk, deep chains, missing auth

### #339 D-2: SQLite Database Backend
- `starforge config db init/migrate/query/backup/export/status/check`
- rusqlite with bundled feature (no system sqlite required)
- Schema: meta, wallets, networks, config_kv, plugins, templates
- TOML import/export for backwards compatibility
- SELECT-only query interface for safe CLI access

## Bug fixes (pre-existing)
- Remove non-existent async-graphql 0.12 / actix-web / tokio deps
- Fix wasm-bindgen-test version (1.3 -> 0.3)
- Fix ureq 2.x resp.ok() usage in registry.rs
- Fix self usage in static methods in test_automation.rs
- Fix use-of-moved-value for TestStatus in test_automation.rs
- Fix duplicate struct fields in test.rs
- Fix multisig_builder.rs import alias and p::step arity
- Fix colored::Colorize on &String (need .as_str())
- Fix social module visibility (load_team, load_review, load_discussion)
- Fix base64::Engine trait not in scope in registry.rs
- Fix collect_files unused root parameter in registry.rs

Closes #329, #330, #331, #339

+++ src/utils/performance.rs
@@ -78,12 +78,9 @@ pub struct GasUsageRecord {
 
 fn metrics_dir() -> Result<PathBuf> {
     let home = dirs::home_dir().ok_or_else(|| anyhow::anyhow!("Could not find home directory"))?;
-    let dir = home
-        .join(".starforge")
-        .join("metrics");
+    let dir = home.join(".starforge").join("metrics");
     if !dir.exists() {
-        fs::create_dir_all(&dir)
-            .with_context(|| format!("Failed to create {}", dir.display()))?;
+        fs::create_dir_all(&dir).with_context(|| format!("Failed to create {}", dir.display()))?;
     }
     Ok(dir)
 }
@@ -191,7 +188,9 @@ pub fn set_alert(
         }
     };
 
-    contract_metrics.alerts.retain(|a| a.metric_name != metric_name);
+    contract_metrics
+        .alerts
+        .retain(|a| a.metric_name != metric_name);
     contract_metrics.alerts.push(AlertConfig {
         metric_name: metric_name.to_string(),
         threshold,
@@ -275,7 +274,11 @@ pub fn generate_report(contract_id: &str, network: &str) -> Result<PerformanceRe
             total_executions: gas_history.len() as u64,
             avg_gas_used: avg_gas,
             max_gas_used: max_gas,
-            min_gas_used: if min_gas == f64::INFINITY { 0.0 } else { min_gas },
+            min_gas_used: if min_gas == f64::INFINITY {
+                0.0
+            } else {
+                min_gas
+            },
             avg_execution_time_ms: avg_time,
             success_rate,
         },
@@ -311,7 +314,7 @@ impl MetricCollector {
         record_gas_usage(&GasUsageRecord {
             contract_id: self.contract_id.clone(),
             operation: "execution".to_string(),
-            gas_used: total_ms as u64 * 100,
+            gas_used: total_ms * 100,
             timestamp: chrono::Utc::now().to_rfc3339(),
             success: true,
             execution_time_ms: total_ms,
